Updates #return_if_errors! to not include errors in exception object:
The rescue for Interrupt merges any errors on the exception into self. This results in duplicate errors on self if we include errors on the exception object. We don't need to merge self's errors into self. They are already present. Problem is they get merged under the :base key, so they are duplicate, but not easily detectable.